RT-125: Убрать отписки, Цветков попросил, добавлена модель юзер и видоплеер
