+++ src/controllers/loginPage.js
@@ -1,6 +1,6 @@
 import { eventBus } from '../modules/eventBus.js';
 import { LogInView } from '../views/LogIn/LogIn.js';
-import { LogInModel } from '../models/logIn.js';
+import { LogInModel } from '../models/loginPage.js';
 import Controller from './controller.js';
 
 /** Class representing login page model. */
