+++ src/controllers/user.js
@@ -1,0 +1,12 @@
+import { eventBus } from '../modules/eventBus.js';
+import Controller from './controller.js';
+import { UserModel } from '../models/user.js';
+
+/** Class representing user controller. */
+export class UserController extends Controller {
+    constructor() {
+        super(new Controller());
+        this.eventBus = eventBus;
+        this.model = new UserModel(this.eventBus);
+    }
+}
