+++ src/models/detailPage.js
@@ -1,11 +1,5 @@
-import {
-    getCurrentUser,
-    getDetailFilm,
-    postAddToFavourites, postDislike,
-    postLike,
-    postRemoveFromFavourites
-} from '../modules/http.js';
-import {Events} from '../consts/events';
+import { getDetailFilm } from '../modules/http.js';
+import { Events } from '../consts/events.js';
 
 /** Class representing detail page about film model. */
 export class DetailPageModel {
@@ -16,10 +10,6 @@ export class DetailPageModel {
     constructor(eventBus) {
         this.eventBus = eventBus;
         this.eventBus.on(Events.DetailPage.GetInfoAboutMovie, this.getInfoAboutFilm);
-        this.eventBus.on(Events.DetailPage.AddToFavourites, this.addToFavourite);
-        this.eventBus.on(Events.DetailPage.RemoveFromFavourites, this.removeFromFavourite);
-        this.eventBus.on(Events.DetailPage.Like, this.like);
-        this.eventBus.on(Events.DetailPage.Dislike, this.dislike);
     }
 
     /**
@@ -38,74 +28,4 @@ export class DetailPageModel {
             this.eventBus.emit(Events.Homepage.Render.ErrorPage);
         });
     };
-
-    /**
-     * Add film to favourites.
-     * @param {string} filmId - Film id, needed to add to favourites.
-     */
-    addToFavourite = (filmId) => {
-        getCurrentUser().then((idUser) => {
-            if (idUser) {
-                postAddToFavourites(filmId).then(() => {
-                    this.eventBus.emit(Events.DetailPage.Change.IconOfFav);
-                });
-            } else {
-                this.eventBus.emit(Events.PathChanged, { path: '/login' });
-            }
-        });
-    }
-
-    /**
-     * Remove film film favourites.
-     * @param {string} filmId - Film id, needed to remove from favourites.
-     */
-    removeFromFavourite = (filmId) => {
-        getCurrentUser().then((idUser) => {
-            if (idUser) {
-                postRemoveFromFavourites(filmId).then(() => {
-                    this.eventBus.emit(Events.DetailPage.Change.IconOfFav);
-                });
-            } else {
-                this.eventBus.emit(Events.PathChanged, { path: '/login' });
-            }
-        });
-    }
-
-    /**
-     * Like film.
-     * @param {string} filmId - Film id, needed to like.
-     */
-    like = (filmId) => {
-        getCurrentUser().then((idUser) => {
-            if (idUser) {
-                postLike(filmId).then(() => {
-                    const data = {
-                        isLike: true,
-                    };
-                    this.eventBus.emit(Events.DetailPage.Change.IconOfLike, data);
-                });
-            } else {
-                this.eventBus.emit(Events.PathChanged, { path: '/login' });
-            }
-        });
-    }
-
-    /**
-     * Dislike film.
-     * @param {string} filmId - Film id, needed to dislike.
-     */
-    dislike = (filmId) => {
-        getCurrentUser().then((idUser) => {
-            if (idUser) {
-                postDislike(filmId).then(() => {
-                    const data = {
-                        isLike: false,
-                    };
-                    this.eventBus.emit(Events.DetailPage.Change.IconOfLike, data);
-                });
-            } else {
-                this.eventBus.emit(Events.PathChanged, { path: '/login' });
-            }
-        });
-    }
 }
