+++ src/models/loginPage.js
@@ -1,0 +1,10 @@
+/** Class representing login page model. */
+export class LogInModel {
+    /**
+     * Create a login page model.
+     * @param {EventBus} eventBus - Global Event Bus.
+     */
+    constructor(eventBus) {
+        this.eventBus = eventBus;
+    }
+}
