+++ src/models/signupPage.js
@@ -1,0 +1,10 @@
+/** Class representing signup page model. */
+export class SignUpModel {
+    /**
+     * Create a signup page model.
+     * @param {EventBus} eventBus - Global Event Bus.
+     */
+    constructor(eventBus) {
+        this.eventBus = eventBus;
+    }
+}
